+++ Api/config/routes.config.js
@@ -1,4 +1,6 @@
 const express = require('express');
+const mongoose = require('mongoose');
+const createError = require('http-errors');
 const router = express.Router();
 const user = require('../controllers/users.controller');
 const groups = require('../controllers/list-groups.controller');
@@ -38,10 +40,48 @@ router.patch('/products/:id', isAuthenticated, isProductOwner, product.update);
 router.delete('/delete-product/:id', isAuthenticated, isProductOwner, product.delete);
 // return service.delete(`/delete-product/${productId}`);
 
-//rutas para el controlador de Store
+// rutas para el controlador de Store
 router.get('/list/:listId/stores', isAuthenticated, storeController.list);
 router.post('/list/:listId/stores', isAuthenticated, storeController.create);
 router.delete('/stores/:storeId', isAuthenticated, storeController.delete);
 router.get('/near', storeController.findNearbyStores);
+
+router.use((req, res, next) => next(createError(404, 'Route not found')));
+
+router.use((error, req, res, next) => {
+
+  if (error.message === 'Not allowed by CORS') {
+    return res.status(403).json({
+      message: 'Cross-Origin Request Blocked: The Same Origin Policy disallows reading the remote resource at ' + req.path,
+      
+      origin: req.get('Origin'),
+      errors: { cors: 'Not allowed by CORS' },
+    });
+  }
+  if(error instanceof mongoose.Error.CastError && error.message.includes('_id')){
+    error = createError(404, 'Resource not found');
+  } else if(error instanceof mongoose.Error.ValidationError){
+      error = createError(400, error);
+  } else if(!error.status) {
+    error = createError(500, error)
+  }
+    // console.log(error);
+    let errors;
+    if(error.errors){
+      errors = Object.keys(error.errors)
+        .reduce((errors, errorKey) => {
+          errors[errorKey] = error.errors[errorKey].message || error.errors[errorKey];
+          return errors;
+      }, {}) 
+    }
+  
+  const data = {
+    message: error.message,
+    errors: errors
+  }
+
+  res.status(error.status).json(data);
+
+})
 
 module.exports = router;
